Add E2E test for vstate transitions
